reorganize the repo; apply code review suggestions
