+++ tsconfig.json
@@ -1,12 +1,13 @@
 {
 	"compilerOptions": {
 		"target": "ES2022",
-		"module": "commonjs",
+		"module": "NodeNext",
+		"moduleResolution": "NodeNext",
 		"esModuleInterop": true,
 		"forceConsistentCasingInFileNames": true,
 		"strict": true,
 		"skipLibCheck": true,
 		"outDir": "./dist"
 	},
-	"include": ["src/**/*"]
+	 "include": ["scripts/**/*"]
 }
